refactor(auto-claude): split utils.ts into focused modules
Split 473-line utils.ts into: shell.ts (exec helpers), claude-cli.ts
(Claude CLI wrappers), labels.ts (GitHub label management), templates.ts
(prompt template resolution). utils.ts retains file helpers and step runner.
