+++ src/commands/auto-claude/status.test.ts
@@ -5,7 +5,7 @@ import { join } from "node:path";
 import { afterAll, beforeAll, describe, expect, it } from "vitest";
 
 import type { Issue } from "../../utils/git/gh-cli-wrapper.js";
-import { LABELS } from "../../lib/auto-claude/utils.js";
+import { LABELS } from "../../lib/auto-claude/labels.js";
 import { checkArtifacts, findAcLabel, formatIssueStatus } from "./status.js";
 
 // ── Test fixtures ──
