+++ src/lib/auto-claude/steps/implement.ts
@@ -4,16 +4,10 @@ import consola from "consola";
 
 import { getConfig } from "../config.js";
 import { ARTIFACTS, STEP_LABELS, TEMPLATES } from "../prompt-templates/index.js";
-import {
-  buildTokens,
-  fileExists,
-  git,
-  log,
-  logStep,
-  readFile,
-  resolveTemplate,
-  runClaude,
-} from "../utils.js";
+import { runClaude } from "../claude-cli.js";
+import { git } from "../shell.js";
+import { resolveTemplate } from "../templates.js";
+import { buildTokens, fileExists, log, logStep, readFile } from "../utils.js";
 import type { IssueContext } from "../utils.js";
 
 export async function stepImplement(ctx: IssueContext): Promise<boolean> {
